+++ dev/app.py
@@ -1,0 +1,18 @@
+from flask import Flask
+from flask import render_template
+from flask_s3 import FlaskS3
+
+
+app = Flask(__name__)
+
+
+@app.route("/")
+def index():
+    return render_template("index.html")
+
+@app.route("/reveal")
+def reveal():
+    return render_template("reveal.html")
+
+if __name__ == "__main__":
+    app.run(host='0.0.0.0',port=5000,debug=True,threaded=True)
