+++ dev/static/lib/custom.css
@@ -1,0 +1,73 @@
+@font-face {
+  font-family: Din-Black;
+  src: url("./fonts/DINBlack.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Plantin;
+  src: url("./fonts/Plantin.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Din-Med;
+  src: url("./fonts/DIN-Medium.ttf") format("truetype");
+}
+
+#canvas {
+  height: 100% ;
+  left: 0px ;
+  position: fixed ;
+  top: 0px ;
+  width: 100% ;
+  z-index: -1 ;
+}
+
+.section p {
+	font-family: "Din-Med", Helvetica,sans-serif;
+	text-align: left; /*justify*/
+	margin: 65px;
+  mix-blend-mode: difference;
+  color: #222a2e;
+}
+
+.keyword {
+  font-family: "Din-Black", Helvetica,sans-serif;
+  line-height: 75%;
+  mix-blend-mode: difference;
+}
+
+.keyword:hover {
+  color: lightgreen;
+}
+
+@media screen and (min-width: 601px) {
+  .section {
+    font-size: 2vw;
+  }
+}
+
+@media screen and (max-width: 600px) {
+  .section {
+    font-size: 4vw;
+  }
+}
+
+a {
+  color: inherit !important; 
+  text-decoration: None !important;
+}
+
+.section {
+        width: 100%; 
+        position: absolute;                      
+}
+
+body {
+  background-color: black !important;
+}
+
+#wrapper {
+  width: 100%;
+  height: 100%;
+  position: relative;
+}
