--- static/lib/custom.css
@@ -1,62 +1,0 @@
-@font-face {
-  font-family: Din-Black;
-  src: url("./fonts/DINBlack.ttf") format("truetype");
-}
-
-@font-face {
-  font-family: Plantin;
-  src: url("./fonts/Plantin.ttf") format("truetype");
-}
-
-
-.section p {
-	font-family: "Plantin", Helvetica,sans-serif;
-	font-size: 44px;
-	text-align: left; /*justify*/
-	margin: 35px; 
-}
-
-.keyword {
-  font-family: "Din-Black", Helvetica,sans-serif;
-  font-size: 44px;
-  
-}
-
-/*span .design {
-  z-index: 5 !important;
-}*/
-
-.section {
-        width: 100%; 
-        z-index:0;
-        position: absolute;                      
-}
-
-#wrapper {
-  width: 100%;
-  height: 100%;
-  position: relative;
-}
-
-.BG {
-  height: 100%;
-  z-index:1;
-  display: inline-block;
-  vertical-align: middle;
-}
-    
-img {
-   position: absolute;
-   top: 24%;
-   left: 25%;
-   width: 50%;
-   display: none;
-}    
-       
-img .app {
-   position: absolute;
-   /*top: 24%;*/
-   left: 25%;
-   /*width: 10%;*/
-   display: none;
-}
