--- templates/index.html
@@ -1,122 +1,0 @@
-<html>
-<head>
-    <title>Carlo Bailey - Designer, New York</title>
-
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel='stylesheet' href='./static/lib/custom.css' />
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-
-</head>
-<body class='' id="body">
-    <!-- <div class="jumbotron" style="background: #FFF"> -->                
-    <div id="wrapper"> 
-		<div class="section">
-			<!-- <div> -->
-				<p>Currently a design researcher at WeWork doing <span class="keyword">spatial analytics</span> and <span id="data-s" class="keyword">data science</span>.
-					Work includes utilizing <span class="keyword">natural language processing</span> to analyze user 
-                    <span id="feeed" class="keyword feedback">feed</span><span id="back" class="keyword feedback">back</span><span id="fData" class="keyword feedback"> data</span>,
-					creating apps to inform 
-                    <span id="d" class="keyword design">d</span><span id="e" class="keyword design">e</span><span id="s" class="keyword design">s</span><span id="i" class="keyword design">i</span><span id="g" class="keyword design">g</span><span id="n" class="keyword design">n</span>
-                     decisions, building predictive <span class="keyword">ML</span> models, <span class="keyword">data
-					visualisation</span> for business intelligence, and sometimes dabbling in 
-                    <span id="I" class="keyword IoT">I</span><span id="o" class="keyword IoT">o</span><span id="T" class="keyword IoT">T</span> 
-                    development.<br><br>
-
-					Went to Columbia University to study <span class="keyword">architecture</span>, focussing on <span class="keyword">design computation</span>,
-					where I also taught a graduate <span class="keyword">class</span> on spatial analytics. Previously worked at Foster + 
-					Partners and was a researcher at KieranTimberlake.<br><br>
-
-					Often speak about <span class="keyword">technology</span> and sometimes write about <span class="keyword">design theory</span>...</p>
-			<!-- </div> -->
-		</div>
-        <div class="BG">
-
-            <!-- DESIGN -->
-            <img id="shang" src="./static/lib/images/WeWork_Weihai-final-large-3.jpg">
-            <img id="cust" src="./static/lib/images/custom.jpg">
-            <img id="berlin" src="./static/lib/images/20160815_Berlin_Sony_Center_-_MVDK-3.jpg">
-            <img id="padd" src="./static/lib/images/20160620_Paddington_London-3.jpg">
-            <img id="wei" src="./static/lib/images/WeWork_Weihai-final-large-8.jpg">
-            <img id="bry" src="./static/lib/images/20141010-8-sm.jpg">
-
-            <!-- IoT -->
-            <img id="iphone1" src="./static/lib/images/OccupancyDshbrd_160525.jpg">
-            <img id="iphone2" src="./static/lib/images/OccupancyDemo_lg-05-160405.jpg">
-            <img id="rasp" src="./static/lib/images/Phonebooth-Sensors.png">
-
-            <!-- FEEDBACK DATA -->
-            <img id="book" class="app" src="./static/lib/images/1_booking.png">
-            <img id="rate" class="app" src="./static/lib/images/2_rating.png">
-            <img id="feed" class="app" src="./static/lib/images/3_feedback.png">
-
-        </div>
-
-    </div>
-
-    <script type="text/javascript" src="https://code.jquery.com/jquery-1.12.4.min.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.11.4/jquery-ui.min.js" type="text/javascript"></script>
-    <script type="text/javascript" src="https://d3js.org/d3.v3.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-    <script type="text/javascript">
-
-    // hide/show images on feedback data
-    $("#feeed").hover(function(){$('#book').show(); },function(){$('#book').hide(); });
-    $("#back").hover(function(){$('#rate').show(); },function(){$('#rate').hide(); });
-    $("#fData").hover(function(){$('#feed').show(); },function(){$('#feed').hide(); });
-
-    // change text colour
-    $('.feedback').hover(function () {
-        $("p").css('color', 'white');
-        $('.keyword').css('color', 'white');
-        $('.feedback').css('color', 'black');
-    },
-    function () {
-        $("p").css('color', 'black');
-        $('.keyword').css('color', 'black');
-    });
-
-
-    // hide/show images on design
-    $("#d").hover(function(){$('#shang').show(); },function(){$('#shang').hide(); });
-    $("#e").hover(function(){$('#cust').show(); },function(){$('#cust').hide(); });
-    $("#s").hover(function(){$('#berlin').show(); },function(){$('#berlin').hide(); });
-    $("#i").hover(function(){$('#padd').show(); },function(){$('#padd').hide(); });
-    $("#g").hover(function(){$('#wei').show(); },function(){$('#wei').hide(); });
-    $("#n").hover(function(){$('#bry').show(); },function(){$('#bry').hide(); });
-
-    // change text colour
-    $('.design').hover(function () {
-        $("p").css('color', 'white');
-        $('.keyword').css('color', 'white');
-        $('.design').css('color', 'black');
-    },
-    function () {
-        $("p").css('color', 'black');
-        $('.keyword').css('color', 'black');
-    });
-
-
-    // hide/show images on IoT
-    $("#I").hover(function(){$('#iphone1').show(); },function(){$('#iphone1').hide(); });
-    $("#o").hover(function(){$('#iphone2').show(); },function(){$('#iphone2').hide(); });
-    $("#T").hover(function(){$('#rasp').show(); },function(){$('#rasp').hide(); });
-
-    // change text colour
-    $('.IoT').hover(function () {
-        $("p").css('color', 'white');
-        $('.keyword').css('color', 'white');
-        $('.IoT').css('color', 'black');
-    },
-    function () {
-        $("p").css('color', 'black');
-        $('.keyword').css('color', 'black');
-    });
-
-
-    </script>
-</body>
-</html>
-
-
