--- templates/index_old.html
@@ -1,207 +1,0 @@
-<html>
-<head>
-    <title>Carlo Bailey - Designer, New York</title>
-
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel='stylesheet' href='./static/lib/custom.css' />
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-
-</head>
-<body class='stop-scrolling' id="body">
-    <!-- <div class="jumbotron" style="background: #FFF"> -->                
-    <div style="display: block; height: 100%;">
-    	<div class="shapes" id="carlo">
-    		<svg id="svgObject" width="300px" height="300px"></svg>             
-        </div>    
-		<div class="section">
-			<div>
-				<p>Carlo Bailey</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>design, research and analytics</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>current work</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>available for collaboration</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>i sometimes write</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>based in new york</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>from london</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>resume here</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>Carlo Bailey</p>
-			</div>
-		</div>
-		<div class="section">
-			<div>
-				<p>design, research and analytics</p>
-			</div>
-		</div>
-
-    </div>
-
-    <script type="text/javascript" src="https://code.jquery.com/jquery-1.12.4.min.js"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.11.4/jquery-ui.min.js" type="text/javascript"></script>
-    <script type="text/javascript" src="https://d3js.org/d3.v3.min.js"></script>
-    <script type="text/javascript" src="./static/lib/snap.svg-min.js"></script>
-    <script type="text/javascript" src="./static/lib/raphael.min.js"></script>
-    <script type="text/javascript" src="./static/lib/animate-object-path.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-    <script type="text/javascript">
-
-
-		var origDocHeight = document.body.offsetHeight*4;
-
-		$(document).ready(function() {
-
-		    function loop() {
-		        $('.section').css({top:0});
-		        $('.section').animate ({
-		            top: '-='+origDocHeight,
-		        }, 50000, 'linear', function() {
-		            loop();
-		        });
-		    }
-		    loop();
-		 });
-
-		var searchDl = 1;
-		var l = 0;
-
-		s = Snap("#svgObject");
-		var path = s.path("M 150, 150 m -50, 0 a 50,50 0 1,0 100,0 a 50,50 0 1,0 -100,0").attr({ fill: "none", stroke: "none"});
-		var r = Raphael(document.getElementById("svgObject"),300, 300);
-
-		var p = r.path("M 150, 150 m -50, 0 a 50,50 0 1,0 100,0 a 50,50 0 1,0 -100,0").attr({stroke: "#FFF", "stroke-width":2}),
-		    pt = path.getPointAtLength(l);
-		    //e = s.circle(160,70,20,20).attr({fill: "white"});
-		    e = r.ellipse(pt.x, pt.y, 10, 10).attr({stroke: "none", fill: "#FFF"}),
-		    totLen = path.getTotalLength()*10,
-		   
-
-		start = function () {
-		    // storing original coordinates
-		    this.ox = this.attr("cx");
-		    this.oy = this.attr("cy");
-		    this.attr({opacity: 1});
-		},
-		move = function (dx, dy) {
-		    var tmpPt = {
-		        x : this.ox + dx, 
-		        y : this.oy + dy
-		    };
-		    // move will be called with dx and dy
-		    l = gradSearch(l, tmpPt);
-		    pt = path.getPointAtLength(l);
-		    this.attr({cx: pt.x, cy: pt.y});
-
-		},
-		up = function () {
-		    // restoring state
-		    this.attr({opacity: 1});
-		},
-		gradSearch = function (l0, pt) {
-		    l0 = l0 + totLen;
-		    var l1 = l0,
-		        dist0 = dist(path.getPointAtLength(l0 % totLen), pt),
-		        dist1,
-		        searchDir;
-		    if (dist(path.getPointAtLength((l0 - searchDl) % totLen), pt) > 
-		       dist(path.getPointAtLength((l0 + searchDl) % totLen), pt)) {
-		        searchDir = searchDl;
-		    } else {
-		        searchDir = -searchDl;
-		    }
-		    console.log(dist1)
-		    l1 += searchDir;
-		    dist1 = dist(path.getPointAtLength(l1 % totLen), pt);
-		    while (dist1 < dist0) {
-		        dist0 = dist1;
-		        l1 += searchDir;
-		        dist1 = dist(path.getPointAtLength(l1 % totLen), pt);
-		    }
-		    l1 -= searchDir;
-		    
-		    return (l1 % totLen);
-		},
-		dist = function (pt1, pt2) {
-		    var dx = pt1.x - pt2.x;
-		    var dy = pt1.y - pt2.y;
-		    return Math.sqrt(dx * dx + dy * dy);
-		};
-		e.drag(move, start, up);
-		e.mousedown($( ".stop-scrolling" ).css({ "overflow": "scroll" }));
-		//e.mouseup($( ".stop-scrolling" ).css({ "overflow": "hidden" }));
-		//e.mousedown($('.section').removeAttr('top'));
-		// function infRotate( el,ref ) {
-		// 	var bbox = ref.getBBox();
-		// 	var trans1 = 'r360,'+(bbox.cx).toString()+','+(bbox.cy).toString();
-		// 	var trans2 = 'r0,'+(bbox.cx).toString()+','+(bbox.cy).toString();
-		//     el.transform(trans1); // removing the previous completed transform.
-		//     el.animate({ transform: trans2 }, -6000, mina.linear, infRotate.bind( null,el,ref ));
-		// };
-
-		//var t = s.ellipse(150, 150, 20, 20).attr({stroke: "none", fill: "#f00"})
-		// infRotate( circle,center )
-		
-		// infRotate( pole, center);
-
-		// // function to pass to svg to enable scrolling
-		// // var test = function(){
-  // //   		console.log("success!");
-  // //   	}
-
-		// //circle.click(test);
-
-		// function drawRect( el ) {
-		// 	el.drawAtPath( path, 6000, { callback: drawRect.bind(null, el) } );
-		// };
-
-		// drawRect(e);
-
-		//////////////////////// Trigonemtry to scroll ////////////////////////
-
-		// var $divs = $('ellipse')
-		// $divs.draggable({
-		//     start: start, cursor: move, stop: up,
-		//     drag: function () {
-		//         var t = $(this), l = $(this).offset().left;
-		//         t.text(l);
-		//         console.log(l)
-		//     }
-		// });
-
-            
-    </script>
-</body>
-</html>
-
-
